refactor(domain): 금액 타입 Long 통일 및 Interest 한국어 레이블 추가
Settlement.sum(BigDecimal), Wallet/WalletTransaction.balance(Integer), Schedule.cost(Integer)를
원화 정수 표현에 맞게 Long으로 통일. Interest에 프론트 표시용 name 필드 추가.

+++ src/main/java/com/buddkitv2/domain/schedule/Schedule.java
@@ -28,7 +28,7 @@ public class Schedule extends BaseEntity {
     @Column(length = 255)
     private String location;
 
-    private Integer cost;
+    private Long cost;
 
     @Enumerated(EnumType.STRING)
     private ScheduleStatus status;
@@ -41,7 +41,7 @@ public class Schedule extends BaseEntity {
     private Club club;
 
     public static Schedule create(String name, LocalDateTime scheduleTime, String location,
-                                   Integer cost, Integer limit, Club club) {
+                                   Long cost, Integer limit, Club club) {
         Schedule s = new Schedule();
         s.name = name;
         s.scheduleTime = scheduleTime;
